Refactored layout of the package
Sequences and mappings have been moved into their own modules. Unit
tests have been refactored in the same way. For the time being names
are exported to the main package, to be readily usable.
