+++ ubimaior/__init__.py
@@ -5,3 +5,13 @@
 __author__ = """Massimiliano Culpo"""
 __email__ = 'massimiliano.culpo@gmail.com'
 __version__ = '0.1.0'
+
+from .sequences import MergedMutableSequence, MergedSequence
+from .mappings import OverridableMapping, MergedMapping
+
+__all__ = [
+    'MergedSequence',
+    'MergedMutableSequence',
+    'MergedMapping',
+    'OverridableMapping'
+]
